ADDED scheduler and timer implementations

+++ occupancy/src/main/kotlin/com/marcusrunge/roominfo/occupancy/interfaces/Scheduler.kt
@@ -1,5 +1,5 @@
 package com.marcusrunge.roominfo.occupancy.interfaces
 
 interface Scheduler {
-    var updateOccupancy: (() -> Unit)?
+    var updateOccupancy: ((Int) -> Unit)?
 }
